Student Information System
Student Information System: It is a student automation system that I wrote in C++. On the console, it has operations such as adding students to the system, removing students from the system, listing students registered in the system. In addition, registered students in the system are also saved in the .txt file format in a formatted way.